♻️ refactor: validation.ts 위치 수정; 불필요한 파일 삭제
