--- Client/src/hooks/HookExample.js
@@ -1,3 +1,0 @@
-const HookExample = () => {};
-
-export default HookExample;
