+++ Client/src/utility/validation.ts
